projects/datasync/pyproject.toml
[project]
name = "datasync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "prefect>=3.3.4",
    "psycopg>=3.2.6",
    "psycopg-binary>=3.2.6",
    "python-dotenv>=1.1.0",
    "questdb>=2.0.3",
    "requests>=2.32.3",
]
